Add IDMS credentials
